# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CreateStudentRequest:
    properties:
      grade:
        maximum: 3
        minimum: 1
        type: integer
      name:
        type: string
      roomNumber:
        type: string
      studentNumber:
        type: string
    required:
    - grade
    - name
    - roomNumber
    - studentNumber
    type: object
...
  dto.PaginatedResponse:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/dto.Pagination'
      success:
        type: boolean
    type: object
  dto.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  dto.Response:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
  dto.StudentResponse:
    properties:
      createdAt:
        type: string
      grade:
        type: integer
      id:
        type: string
      name:
        type: string
      roomNumber:
        type: string
      studentNumber:
        type: string
    type: object
...
  dto.UpdateStudentRequest:
    properties:
      grade:
        maximum: 3
        minimum: 1
        type: integer
      name:
        type: string
      roomNumber:
        type: string
      studentNumber:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 기숙사 관리 시스템 API - 학생 상벌점 관리 및 당직 일정 운영
  title: Dormi API
  version: "1.0"
paths:
...
  /students:
    get:
      description: 학생 목록 조회 (검색, 필터링 지원)
      parameters:
      - description: 검색어 (이름, 학번)
        in: query
        name: search
        type: string
      - description: 학년
        in: query
        name: grade
        type: integer
      - description: 방 번호
        in: query
        name: room
        type: string
      - default: 1
        description: 페이지
        in: query
        name: page
        type: integer
      - default: 20
        description: 페이지당 개수
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.StudentResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 학생 목록 조회
      tags:
      - 학생
    post:
      consumes:
      - application/json
      description: 새로운 학생 등록
      parameters:
      - description: 학생 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateStudentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.StudentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 학생 생성
      tags:
      - 학생
  /students/{id}:
    delete:
      description: 학생 정보 삭제 (soft delete)
      parameters:
      - description: 학생 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 학생 삭제
      tags:
      - 학생
    get:
      description: ID로 학생 정보 조회
      parameters:
      - description: 학생 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.StudentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 학생 상세 조회
      tags:
      - 학생
    put:
      consumes:
      - application/json
      description: 학생 정보 수정
      parameters:
      - description: 학생 ID
        in: path
        name: id
        required: true
        type: string
      - description: 수정할 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateStudentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.StudentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 학생 수정
      tags:
      - 학생
  /students/import:
    post:
      consumes:
      - multipart/form-data
      description: CSV 파일로 학생 일괄 등록
      parameters:
      - description: CSV 파일
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: CSV 일괄 등록
      tags:
      - 학생
...
securityDefinitions:
  BearerAuth:
    description: JWT 토큰 (Bearer {token})
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"